centre section of home completed
